[ADD] generate mip map func in resource update and add move ptr for upload operation

+++ Engine/Source/Runtime/Core/Include/LowRenderer/RhiTypedef.h
@@ -252,7 +252,7 @@ BEGIN_PCCORE
 
 
 #pragma region GetBytePerPixel
-    static inline uint32_t GetBytePerPixel(RhiFormat RhiFormat)
+    static inline size_t GetBytePerPixel(RhiFormat RhiFormat)
     {
         switch (RhiFormat)
         {
